+++ nflow-engine/src/test/java/io/nflow/engine/internal/workflow/StateExecutionImplTest.java
@@ -12,6 +12,7 @@ import static org.hamcrest.Matchers.empty;
 import static org.hamcrest.Matchers.emptyCollectionOf;
 import static org.hamcrest.Matchers.hasEntry;
 import static org.joda.time.DateTime.now;
+import static org.junit.Assert.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
@@ -317,6 +318,43 @@ public class StateExecutionImplTest {
         .setBusinessKey("business").build();
     createExecution();
     execution.handleRetryAfter(tomorrow, def);
+  }
+
+  @Test
+  public void handleFailureGoesToFailureState() {
+    handleFailure(TestDefinition.TestState.start1, TestDefinition.TestState.start1);
+
+    assertFalse(execution.isRetry());
+    assertThat(execution.getNextState(), is(equalTo(TestDefinition.TestState.failed.name())));
+    assertThat(execution.getNextActivation(), is(notNullValue()));
+    assertThat(execution.getNextStateReason(), is("reason, going to failure state"));
+  }
+
+  @Test
+  public void handleFailureGoesToErrorStateWhenFailureStateIsNotDefined() {
+    handleFailure(TestDefinition.TestState.start2, TestDefinition.TestState.start2);
+
+    assertFalse(execution.isRetry());
+    assertThat(execution.getNextState(), is(equalTo(TestDefinition.TestState.error.name())));
+    assertThat(execution.getNextActivation(), is(notNullValue()));
+    assertThat(execution.getNextStateReason(), is("reason, no failure state defined, going to error state"));
+  }
+
+  @Test
+  public void handleFailureStopsProcessingWhenProcessingErrorState() {
+    handleFailure(TestDefinition.TestState.start2, TestDefinition.TestState.error);
+
+    assertFalse(execution.isRetry());
+    assertThat(execution.getNextState(), is(equalTo(TestDefinition.TestState.error.name())));
+    assertThat(execution.getNextActivation(), is(nullValue()));
+    assertThat(execution.getNextStateReason(), is("reason when handling error state, processing stopped"));
+  }
+
+  private void handleFailure(TestDefinition.TestState initialState, TestDefinition.TestState currentState) {
+    TestDefinition def = new TestDefinition("x", initialState);
+    instance = new WorkflowInstance.Builder().setState(currentState.name()).build();
+    createExecution();
+    execution.handleFailure(def, "reason");
   }
 
   static class Data {