+++ nflow-tests/src/main/java/io/nflow/tests/demo/workflow/FibonacciWorkflow.java
@@ -46,11 +46,6 @@ public class FibonacciWorkflow extends WorkflowDefinition<FibonacciWorkflow.Stat
     public WorkflowStateType getType() {
       return type;
     }
-
-    @Override
-    public String getDescription() {
-      return name();
-    }
   }
 
   public FibonacciWorkflow() {